chore: DRY up prediction logic
Prompt:
Right now, the Perceptron prediction is repeated in several places. Please dry that up and make sure we only have a single Perceptron instance that gets updated whenever the dials are changed of W1, W2, or the bias. And make sure that that single instance of a Perceptron is used for all predictions. For the color of the prediction, in the test classification area, for the result, and so on. We don't want this logic to exist in multiple places, but just in a single place.